// src/app/models/laptop.ts
export interface Laptop {
  id: string;
  name: string;
  picture: string;
  productCode: string;
  price: number;
  productType: string;
  productCreator: string;
  series: string;
  model: string;
  computerType: string;
  weight: string;
  length: string;
  screenSize: string;
  screenResolution: string;
  operationalSystem: string;
  processorClass: string;
  processorType: string;
  ramType: string;
  amountOfRam: string;
  maxAmountOfRam: string;
  dataStorageType: string;
  dataStorageSize: string;
  graphicsCardType: string;
  graphicsCardSeries: string;
  graphicsCardModel: string;
  graphicsCardMemorySize: string;
  videoCamera: boolean;
  integratedMicrophone: boolean;
  keyboardLanguage: string;
  category: string;
  computerAndOfficeCategory: string;
}